requirements.txt
Flask==2.0.1
Flask-OAuthlib==0.9.6
python-dotenv==1.0.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
oauthlib==2.1.0
